feat(domain) : Lotto 클래스 및 유효성 테스트 추가
