+++ src/test/java/lotto/domain/LottoTest.java
@@ -1,0 +1,46 @@
+package lotto.domain;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+import java.util.List;
+
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
+
+class LottoTest {
+    @Test
+    void 로또_번호의_개수가_6개가_넘어가면_예외가_발생한다() {
+        assertThatThrownBy(() -> new Lotto(List.of(1, 2, 3, 4, 5, 6, 7)))
+                .isInstanceOf(IllegalArgumentException.class);
+    }
+
+    @DisplayName("로또 번호에 중복된 숫자가 있으면 예외가 발생한다.")
+    @Test
+    void 로또_번호에_중복된_숫자가_있으면_예외가_발생한다() {
+        assertThatThrownBy(() -> new Lotto(List.of(1, 2, 3, 4, 5, 5)))
+                .isInstanceOf(IllegalArgumentException.class);
+    }
+
+    @DisplayName("로또 번호가 1부터 45 사이가 아니면 예외가 발생한다.")
+    @Test
+    void 로또_번호가_범위를_벗어나면_예외가_발생한다() {
+        assertThatThrownBy(() -> new Lotto(List.of(0, 2, 3, 4, 5, 6)))
+                .isInstanceOf(IllegalArgumentException.class)
+                .hasMessageContaining("1부터 45");
+    }
+
+    @DisplayName("로또 번호가 올바를 경우 정상적으로 생성된다.")
+    @Test
+    void 로또_번호가_정상적으로_생성된다() {
+        Lotto lotto = new Lotto(List.of(1, 2, 3, 4, 5, 6));
+        assertThat(lotto).isNotNull();
+    }
+
+    @DisplayName("로또 번호는 오름차순으로 정렬된다.")
+    @Test
+    void 로또_번호는_오름차순으로_정렬된다() {
+        Lotto lotto = new Lotto(List.of(10, 1, 8, 5, 3, 7));
+        assertThat(lotto.getNumbers()).containsExactly(1, 3, 5, 7, 8, 10);
+    }
+}
